% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trust_data.R
\name{academies.trust.data}
\alias{academies.trust.data}
\title{Import and Sequence Academy Trust History Data}
\usage{
academies.trust.data(gias_date, cut_off_date, urn_link_type = "ofsted")
}
\arguments{
\item{gias_date}{the file date of the GIAS cut, defaults to first day of current month if not entered.}

\item{cut_off_date}{the cut off date for open schools, defaults to first day of current month if not entered.}

\item{urn_link_type}{defaults to Ofsted URN linking methodology, use "all" if merges are to be included. Note this will likely flag a warning for multiple join dates.}
}
\value{
Dataframe of academies and trust history.
}
\description{
\code{academies.trust.data()} imports GIAS Academies Trust Membership History Data from https://www.get-information-schools.service.gov.uk/Downloads. Standardises variable names to be unique and consist only of characters, numbers, and letters and sets all date and numeric variables to relevant classes. Removes duplicates and adds a history_order variable to show 1 = current trust.
}
\details{
Returns the following variables in a df:
\itemize{
\item \code{current_urn} The current URN of the schools using either the Ofsted methodology or links including all merges (no splits)
\item \code{group_id} The Group ID of the Trust
\item \code{type_of_establishment_name} Shows whether the school is sponsored academy or an academy converter according to the status when the school first academised
\item \code{date_joined_group}
\item \code{date_left_group}
\item \code{history_order} The order in which the school joined the trust with 1 = most recent/current
\item \code{n_trusts} The number of trusts the school has been a member of
\item \code{group_type} Whether the trust is a Single or Multi-Academy Trust (SAT or MAT)
}
}
\examples{
academies.trust.data("2021-06-01", "2021-06-01", urn_link_type = "all")

}
