% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academies.R
\name{academy.pipeline}
\alias{academy.pipeline}
\title{Import Academy Pipeline Data}
\usage{
academy.pipeline(converter = TRUE, sponsored = TRUE)
}
\arguments{
\item{converter}{TRUE includes academy converter data in the output df}

\item{sponsored}{TRUE includes sponsored academy data in the output df}
}
\value{
Dataframe of pipeline data.
}
\description{
\code{academy.pipeline()} returns a df containing basic data on the converter and/or sponsored academy pipelines from the latest pipeline MI: https://www.gov.uk/government/publications/open-academies-and-academy-projects-in-development
}
\details{
Returns the following variables in a df:
\itemize{
\item \code{urn} the URN of the school as listed in the pipeline MI datacurrent_pipeline_sponsored,
\item \code{application_date} for sponsored academies this variable reflects the data the project was approved (dAO issue date), for academy converters this reflects the date the application was approved.
\item \code{name_of_matched_sponsor} populated only for sponsored academies,
\item \code{proposed_opening_date} proposed date for the academy to open,
\item \code{academy_type} indicates whether the academy is a sponsored academy or academy converter
}
}
\examples{
academy.pipeline(converter = TRUE, sponsored = FALSE)

}
