% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gias_downloads.R
\name{gias.estab.fields}
\alias{gias.estab.fields}
\title{Import GIAS Establishment Fields Data}
\usage{
gias.estab.fields(gias_date)
}
\arguments{
\item{gias_date}{the file date of the GIAS cut, defaults to first day of current month if not entered.}
}
\value{
Dataframe of GIAS establishment fields data with standardised names and variable class assigned.
}
\description{
Imports GIAS Establishment Fields Data from https://www.get-information-schools.service.gov.uk/Downloads.
Standardises variable names to be unique and consist only of the _ character, numbers, and letters and sets all date and numeric variables to relevant classes.
}
\examples{
gias.estab.fields("2021-06-01")

}
