% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giasr-package.R
\docType{package}
\name{giasr-package}
\alias{giasr}
\alias{giasr-package}
\title{giasr: GIAS Data Wrangling Functions}
\description{
A set of functions for cleaning and shaping GIAS data.
}
\author{
\strong{Maintainer}: Pennie Vargas \email{pennie.vargas@education.gov.uk}

}
\keyword{internal}
