% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographies.R
\name{lad.lookup}
\alias{lad.lookup}
\alias{priority.areas}
\title{Get Establishment Location Data}
\usage{
lad.lookup(gias_date, schools_open_date)

priority.areas(gias_date, schools_open_date)
}
\arguments{
\item{gias_date}{the file date of the GIAS cut, defaults to first day of current month if not entered.}

\item{schools_open_date}{the date on which the schools were open, defaults to returning data for all establishments when not provided}
}
\value{
Dataframe of schools with 2018, 2019, and 2020 LAD locations.

\code{priority.areas} includes OA and ONE LADs where applicable
}
\description{
\code{lad.lookup} uses eastings and northings data from GIAS and LAD boundary data from the ONS to identify historic LAD locations of schools from 2018.
\code{priority.areas} uses eastings and northings data from GIAS and LAD LAD boundary data from the ONS to identify schools in target areas for ONE and OA place-based programmes.
}
\details{
NOTE: due to the size of the datasets and amount of processing required to output the dataframe this function may take some time to run.
}
\section{Functions}{
\itemize{
\item \code{priority.areas()}: provides a df of all state funded establishments with 2018, 2019, and 2020 LAD locations and OA and/or ONE area membership.

}}
\examples{
lad.lookup("2021-06-01", "2021-06-01")
priority.areas("2021-06-01", "2021-06-01")

}
