% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_state_schools.R
\name{state.schools}
\alias{state.schools}
\alias{current.state.schools}
\title{Identify State Funded Establishments}
\usage{
state.schools(gias_date)

current.state.schools(gias_date, open_date)
}
\arguments{
\item{gias_date}{default is to get GIAS data from the 1st of the current month. However, to reproduce historic analysis or to use more recent data the date can be set using the "\%Y-\%m-\%d" format.}

\item{open_date}{default is to use the 1st of the current month for open schools. However, to reproduce historic analysis or identify more recently open schools the date can be set using the "\%Y-\%m-\%d" format.}
}
\value{
A list of state funded establishment URNs along with
\itemize{
\item \code{phase_type_grouping}: State-funded Nursery, State-funded Primary, State-funded Secondary, State-funded Special school, Pupil referral unit
\item \code{general_type_group}: Academies, Local Authority Maintained
\item \code{type_of_establishment_name}: GIAS fields
\item \code{phase_of_education_name}: GIAS fields
}
}
\description{
The \code{state.schools()} function is used to identify a list of all schools that are state funded establishments in the GIAS all establishment data released on the selected date.
While \code{current.state.schools()} identifies all currently open state funded establishments as at a specified date, using the GIAS all establishment data released on the selected date.
}
\details{
Schools identified as state funded establishments align with those in external reporting. These are:
\itemize{
\item Academy 16 to 19 sponsor led
\item Academy 16-19 converter
\item Academy alternative provision converter
\item Academy alternative provision sponsor led
\item Academy converter
\item Academy special converter
\item Academy special sponsor led
\item Academy sponsor led
\item City technology college
\item Community school
\item Community special school
\item Foundation school
\item Foundation special school
\item Free schools
\item Free schools 16 to 19
\item Free schools alternative provision
\item Free schools special
\item Local authority nursery school
\item Pupil referral unit
\item Studio schools
\item University technical college
\item Voluntary aided school
\item Voluntary controlled school
}
}
\section{Functions}{
\itemize{
\item \code{state.schools}: Identifies all state funded establishments (historc and currently open) as at a specified date, using the GIAS all establishment data released on the selected date.

\item \code{current.state.schools}: Identifies all currently open state funded establishments as at a specified date, using the GIAS all establishment data released on the selected date.
}}

\examples{
current.state.schools("2021-06-01", "2021-06-01")
state.schools("2021-06-01")

}
