% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urn_linking.R
\name{urn.links.no.splits}
\alias{urn.links.no.splits}
\alias{ofsted.urn.links}
\title{Identify URN Links}
\usage{
urn.links.no.splits(gias_date, cut_off_date)

ofsted.urn.links(gias_date, cut_off_date)

ofsted.urn.links(gias_date, cut_off_date)
}
\arguments{
\item{gias_date}{default is to get GIAS data from the 1st of the current month. However, to reproduce historic analysis or to use more recent data the date can be set using the "\%Y-\%m-\%d" format.}

\item{cut_off_date}{default is to use the 1st of the current month for open schools. However, to reproduce historic analysis or identify more recently open schools the date can be set using the "\%Y-\%m-\%d" format.}
}
\value{
A df of \code{urn} and \code{current_urn} identifiers. This can be used to link data from other sources (e.g. school census, performance data)
\itemize{
\item \code{urn}: All past and present URNs
\item \code{current_urn}: The most recent URN relating to the URN shown in the \code{urn} variable. This may be the same as the \code{urn}, or a successor. \code{current_urn} establishments may be closed.
}
}
\description{
The \code{urn.links.no.splits()} function is used to identify all state funded establishments that link to the most recent single URN for a state funded establishment, excluding splits. \cr
While \code{ofsted.urn.links()} function identifies only those links for state funded establishments that would be considered links using the Ofsted methodology.
}
\details{
The resulting df for \code{urn.links.no.splits} can be used to link and aggregate data from other sources (e.g. school census, performance data). While the resulting df from \code{ofsted.urn.links} should be used primarily to link Ofsted data to the current school.
}
\section{Functions}{
\itemize{
\item \code{ofsted.urn.links}: Identifies predecessor-successor links following the Ofsted methodology as at a specified date, using the GIAS all establishment data released on the date provided and open on the date provided.
}}

\examples{
urn.links.no.splits("2021-06-01", "2021-06-01")
ofsted.urn.links("2021-06-01", "2021-06-01")

}
